Use libled from ledmon for get/set SES LEDs
Replace libstoragemgmts implementation of SES for what is including
it libled.  We also pickup NPEM & VMD support in ledmon.  As
libstoragemgmt functionality includes "getting" current values we
can't leverage libled support of additional hardware that allows
you to set LEDs, but not query.

We are leveraging the slots API in libled to achieve get/set
functionality.  Moving forward we can also expose the libled
slots API in libstoragemgmt so that users can get/set LEDs when
a device node isn't accessible.

Todo:
- Write a common function to translate libled error code to appropriate
  libstoragemgmt error code with associated error text.

Signed-off-by: Tony Asleson <tasleson@redhat.com>