megaraid: Handle case where no drives exist
When we simply have an adapter with nothing connected to
it, make sure we return an empty hashtable instead of None.
This is because callers assume they can call .get() on
whatever is returned, which doesn't work for None.

Note: The command executes fine, but there is no "Response Data"
in the json, eg.

{
  "Controllers": [
    {
      "Command Status": {
        "CLI Version": "007.1616.0000.0000 Dec 24, 2020",
        "Operating system": "Linux 6.8.6-200.fc39.x86_64",
        "Status Code": 0,
        "Status": "Success",
        "Description": "None"
      },
      "Response Data": {
        "Controller Count": 1
      }
    }
  ]
}

vs.

{
  "Controllers": [
    {
      "Command Status": {
        "CLI Version": "007.1616.0000.0000 Dec 24, 2020",
        "Operating system": "Linux 6.8.6-200.fc39.x86_64",
        "Controller": 0,
        "Status": "Success",
        "Description": "Show Drive Group Succeeded"
      }
    }
  ]
}

Signed-off-by: Tony Asleson <tasleson@redhat.com>