Suppress compiler warning
Modern versions of clang will flag the var_arg parameter if it's
a promotable type (like an unscoped enum), the standard declares
the whole thing undefined behavior.  We're suppressing this warning, and
the code does indeed work for both gcc and clang compilers.

The way to fix this correctly, is to change the function signature to:

int lsm_capability_set_n(lsm_storage_capabilities *cap,
                          int v, ...);

That would support binary ABI compatibility, but would cause a source
level change.  That change is still up for consideration.

Signed-off-by: Tony Asleson <tasleson@redhat.com>